length fix

+++ VinFletcherArrow/Arrow.cs
@@ -18,16 +18,16 @@
     {
         public ArrowheadType arrowHead;
         public Fletching fletching;
-        public float length;
+        public int length;
 
-        public Arrow(int arrowHead, int fletching, float length)
+        public Arrow(int arrowHead, int fletching, int length)
         {
             this.arrowHead = (ArrowheadType)arrowHead;
             this.fletching = (Fletching)fletching;
             this.length = length;
         }
 
-        public Arrow (ArrowheadType arrowHead, Fletching fletching, float length)
+        public Arrow (ArrowheadType arrowHead, Fletching fletching, int length)
         {
             this.arrowHead = arrowHead;
             this.fletching = fletching;
